Detail invoice penjualan

+++ application/controllers/Dashboard.php
@@ -51,12 +51,18 @@
 
     public function proses_pesanan()
     {
-      
+      $is_processed = $this->model_invoice->index();
+      if($is_processed){
         $this->cart->destroy();
         $this->load->view('templates/header');
         $this->load->view('templates/sidebar');
         $this->load->view('proses_pesanan');
         $this->load->view('templates/footer');
+
+      }else{
+        echo "Maaf, Pesanan anda gagal diproses";
+      }
+
     }
 
   }
